Refactor into <CodapDndContext>
